Move imports to tableContents.js
