+++ frontend/src/worksheets/workSheetPDF.js
@@ -1,41 +1,28 @@
 import pdfMake from 'pdfmake/build/pdfmake'
 import vfsFonts from 'pdfmake/build/vfs_fonts'
-import {
-  typeOfWorkTranslation,
-  assetSettlement,
-  typeOfPayment,
-} from './TranslationForWorkSheet'
-import renderSvg from './renderSvg'
+
 import {
   createHeader,
   createDetails,
   createDescription,
   createMaterials,
   createSignatureAndDate,
-  createSignatureAndDate,
+  createStyles,
 } from './tableContents'
 
 function workSheetPDF(worksheet) {
-  const workerSignatureSvg = renderSvg(worksheet.workerSignature)
-  const proofOfEmploymentSvg = renderSvg(worksheet.proofOfEmployment)
-
   const { vfs } = vfsFonts.pdfMake
   pdfMake.vfs = vfs
 
-  var doc = {
+  const doc = {
     content: [
-      {
-        text: 'Munkalap',
-        style: 'header',
-        alignment: 'right',
-      },
       createHeader(worksheet),
       createDetails(worksheet),
       createDescription(worksheet),
       createMaterials(worksheet),
       createSignatureAndDate(worksheet),
-      ],
-      createStyles()
+    ],
+    style: createStyles(),
   }
 
   pdfMake.createPdf(doc).open()
